+++ files/CHANGELOG.md
@@ -8,6 +8,8 @@ Also with this versions, typings tests are provided and several definitions are 
 
 - Deprecate `R.contains`
 
+- Deprecate `R.defaultToWhen`
+
 2.17.0 Big change in `R.runTests` logic. It will be removed from Rambdax to `helpers` repo.
 
 2.16.0 Restore `R.runTests` but without documentation